requirements.txt
# Indian Sign Language Detection - Dependencies
# Install with: pip install -r requirements.txt

# Core ML Libraries
tensorflow>=2.13.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Computer Vision
opencv-python>=4.8.0
mediapipe>=0.10.0

# Data Handling
Pillow>=10.0.0
matplotlib>=3.7.0

# Dataset Download
kagglehub>=0.2.0
